refactor: Migrate orbit-exec tests to per-module tests/ layout [ORB-00226]
